convertendo para tipos genericos

+++ ZCRQS/Core.Shared/DataWithVersion.cs
@@ -2,13 +2,13 @@ namespace Core.Shared
 {
     public class DataWithVersion
     {
-        public DataWithVersion(int version, byte[] data)
+        public DataWithVersion(long version, byte[] data)
         {
             Version = version;
             Data = data;
         }
 
-        public int Version { get; private set; }
+        public long Version { get; private set; }
         public byte[] Data{ get; private set; }
     }
 }
